requirements.txt
Flask==1.0.2
pymongo==3.6.0
mongoengine==0.15.0
flask-mongoengine==0.9.5
python-dotenv==0.7.1
pytest==3.5.0
factory_boy
